@@ -4,15 +4,15 @@
   "private": true,
   "dependencies": {
     "babel-runtime": "^6.26.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^16.10.2",
+    "react-dom": "^16.10.2",
     "react-loadable": "^5.5.0",
-    "react-router-dom": "^5.0.1",
+    "react-router-dom": "^5.1.2",
     "semantic-ui-react": "^0.87.3"
   },
   "devDependencies": {
     "babel-core": "^6.26.3",
-    "babel-eslint": "^10.0.2",
+    "babel-eslint": "^10.0.3",
     "babel-loader": "7",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-env": "^1.7.0",
@@ -21,24 +21,24 @@
     "css-loader": "^2.1.1",
     "eslint": "5.3.0",
     "eslint-config-airbnb": "17.1.1",
-    "eslint-config-prettier": "^6.0.0",
+    "eslint-config-prettier": "^6.4.0",
     "eslint-loader": "^2.2.1",
     "eslint-plugin-import": "2.18.0",
     "eslint-plugin-jsx-a11y": "6.2.3",
-    "eslint-plugin-prettier": "^3.1.0",
+    "eslint-plugin-prettier": "^3.1.1",
     "eslint-plugin-react": "7.14.2",
-    "file-loader": "^4.0.0",
+    "file-loader": "^4.2.0",
     "html-webpack-plugin": "^3.2.0",
     "node-sass": "^4.12.0",
     "postcss-loader": "^3.0.0",
     "prettier": "^1.18.2",
-    "sass-loader": "^7.1.0",
+    "sass-loader": "^7.3.1",
     "style-loader": "^0.23.1",
-    "svg-url-loader": "^3.0.0",
+    "svg-url-loader": "^3.0.2",
     "url-loader": "^1.1.2",
-    "webpack": "^4.32.2",
-    "webpack-cli": "^3.3.2",
-    "webpack-dev-server": "^3.4.1"
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.9",
+    "webpack-dev-server": "^3.8.2"
   },
   "scripts": {
     "dev": "webpack-dev-server --env.dev",
